==> docs/_sources/notebooks/class_5/week_4/11_part1_emissions.ipynb <==
# Preparing the data: Carbon dioxide emissions

We're walking through our four datasets, processing one at a time:

1. **Carbon dioxide emissions by country**
2. Income (as measured by GDP per capita) by country
3. Population by country (so we can convert CO2 emissions into per capita emissions)
4. List of territories by continent (since we want to be able to group the countries by region of the world)

Let's start by loading our data into a pandas DataFrame and take a look at the first 10 rows of content:

import pandas as pd

co2_all = pd.read_csv("data/raw/co2_emissions_tonnes_per_person.csv")
co2_all.head(10)

Let's also use the `pandas` `describe()` method to summarize the contents:

co2_all.describe()

Looking at the data we learn a few things. First, the dataset contains data on CO2 starting in 1799 and going through 2017. For the plot we will be creating, we only need to plot one year of data, likely the most recent year. Since we only have data up to 2017, we'll need to use that as the common year. Going forward, we'll only concern ourselves with data from 2017.

Second, we see that many of those early years have `NaN` values, which indicates there is no data present. We can see in the "count" row of the description above that in 1799 and 1800, only 5 entries are non-empty, and this number is 194 from 2008 forward. Let's see how many countries we have in total, by using the `shape` attribute of `co2_all`:

co2_all.shape

We have 194 rows and 194 non-empty values in the later years that are included, so we have one value for each country on this list in 2017. We'll want to use the latest data that we have to make the plot as relevant as possible, so the last thing to do is to just extract the 2017 data. We'll need the 2017 column for the data as well as the corresponding country/territory so we can combine the data later with income and population:

co2 = co2_all[["country", "2017"]]
co2

One last step: since we'll eventually be merging our columns together, let's rename the '2017' column so that it is more self-explanatory and call it 'co2' using the `rename()` method:

co2 = co2.rename(columns={"2017": "co2"})
co2

Lastly, let's save our data to file.

co2.to_csv('data/intermediate/co2.csv', index=False)

With all of this complete, let's create a function that includes each of the steps above to reproduce what we've done:

import pandas as pd

# Function to process CO2 data.
def process_co2(infile,outfile):
    # Read the data
    co2_all = pd.read_csv(infile)
    
    # Select the columns that we need and rename them for ease of reference
    co2 = co2_all[["country", "2017"]]
    co2 = co2.rename(columns={"2017": "co2"})
    
    # Output the resulting data
    co2.to_csv(outfile, index=False)

# Test to make sure the results are the same as the original
raw_co2_file = "data/raw/co2_emissions_tonnes_per_person.csv"
co2_file = 'data/intermediate/co2.csv'
process_co2(raw_co2_file,co2_file)

The CO2 emissions dataset is ready for use! Let's move on to exploring income per person.
