==> docs/_sources/notebooks/class_5/week_4/12_part2_income.ipynb <==
# Preparing the data: Income

We're walking through our four datasets, processing one at a time:

1. *Carbon dioxide emissions by country*
2. **Income (as measured by GDP per capita) by country**
3. Population by country (so we can convert CO2 emissions into per capita emissions)
4. List of territories by continent (since we want to be able to group the countries by region of the world)

Again, let's start by loading and viewing the first 10 rows of our data:

import pandas as pd

income_all = pd.read_csv("data/raw/income_per_person_gdppercapita_ppp_inflation_adjusted.csv")
income_all.head(10)

As we look over the dataset here, we can see we have data that goes from 1799 through 2049. So clearly the later years are projections rather than measured data (unless the data providers have a time traveler on staff!).

We know that we're only concerned with the year 2017 from our discussion of CO2 data, so let's extract the year 2017 and the country:

income_2017 = income_all[["country", "2017"]]
income_2017


But notice something odd here... Some of the values have the letter 'k' at the end of a number. While 'k' is the SI shorthand for a thousand units (representing "kilo"). But what is it doing in our dataset? Let's take a look at the data types in our data. We'd expect those to be numbers: floats or integers or something similar.

income_2017.dtypes

OK, the data from 2017 are objects, not floats or integers, just like the column for 'country'. Let's take a look at one of the entries in 2017 and see what's going on with the data there.

first_entry = income_2017["2017"][0]
first_entry

type(first_entry)

It's a string! It's not a number. The authors of the data represented the number 12,000 as the string '12k' instead. We need to correct this since all data that we plot will need to be numerical.

For each value, we'll need to read in the string - if there are no letters in the entry, we just need to convert it to a number. If there is a letter k in it, we need to remove that letter and THEN convert the rest into a number and multiply it by 1,000. Let's create a function that does that:

We'll write a function that takes a string as input and outputs the correct numerical representation of the number as a float. That is, if the input is '20' then the output should be the number 20.0; if the input is '20k' then the output should be the number 20000.0

def string2num(string):
    if "k" in string:
        number = float(string[:-1]) * 1000
    else:
        number = float(string)
    return number


print(string2num("2"))
print(string2num("2k"))

Now let's apply our function to our data for each entry in 2017. But before we do, let's make a copy of our data so we don't edit the original (`deep = True`) ensures all the data are a copy rather than a view):

income = income_2017.copy(deep=True)

We can use the `pandas` method `apply()` to apply our `string2num` function to *each entry* in our '2017' column:

income["2017"] = income_2017["2017"].apply(string2num)
income

While it looks like we have successfully converted the strings to floats, let's look at our data types again to ensure we fixed the issue:

income.dtypes

Great! We have a column of floats now, so it looks like our fix worked. Let's make one more check to see whether there are any `NaN` values introduced - we want to avoid introducing anything like that into the mix:

income["2017"].isna().values.any()

No `NaN` values - so we're all set! 

And as our last step, let's also rename the column header from '2017' to 'income', similar to what we did with `co2` above, so that its content is clear and it will be ready to be merged together after we've loaded each of our four datasets.

income = income.rename(columns={"2017": "income"})
income

Great - this looks ready to go! Let's save this to file.

income.to_csv('data/intermediate/income.csv', index=False)

Wonderful! With that complete, let's again collect our steps and various functions into a single function that we can run for reproducibility:

import pandas as pd

# Function to process income data
def process_income(infile,outfile):
    # Helper function to convert string data within the file into numerical data
    def string2num(string):
        if "k" in string:
            number = float(string[:-1]) * 1000
        else:
            number = float(string)
        return number
    
    # Read the file, select the columns of interest
    income_all = pd.read_csv(infile)
    income_2017 = income_all[["country", "2017"]]
    
    # Convert the textual data in the dataset into numerical data
    income = income_2017.copy(deep=True)
    income["2017"] = income_2017["2017"].apply(string2num)
    
    # Rename the columns to make them more readable and save the output to file
    income = income.rename(columns={"2017": "income"})
    income.to_csv(outfile, index=False)

# Test and make sure the file produces identical output as the original process:
raw_income_file = "data/raw/income_per_person_gdppercapita_ppp_inflation_adjusted.csv"
income_file = "data/intermediate/income.csv"
process_income(raw_income_file,income_file)

With both CO2 and income data prepared, population is next!
