==> docs/_sources/notebooks/class_5/week_4/13_part3_population.ipynb <==
# Preparing the data: Income

We're walking through our four datasets, processing one at a time:

1. *Carbon dioxide emissions by country*
2. *Income (as measured by GDP per capita) by country*
3. **Population by country (so we can convert CO2 emissions into per capita emissions)**
4. List of territories by continent (since we want to be able to group the countries by region of the world)

As is our standard practice at this point, let's load in and view the first 10 entries in our dataset:

import pandas as pd
pop_all = pd.read_csv("data/raw/population_total.csv")
pop_all.head(10)

These data go well into the future, all the way out to 2099, so again, these are projections. Let's begin by grabbing the 2017 data and dive a bit deeper.

pop_2017 = pop_all[["country", "2017"]]
pop_2017

It looks like we have suffixes of 'k' and 'M' (representing millions). We will have to apply a similar function to what we used for income here to prepare these data. But we have to ask the question: are 'k' and 'M' the *only* suffixes included in the data or are there others? Let's write a function to figure that out.

We'll create a function that takes in a `numpy` array (into which we will pass the 2017 population values of `pop_2017['2017'].values`) and searches through them to find any alphabetic letters that are present. We'll also make use of the string method `isalpha()` and the `numpy` method `unique()` to help with this process.

For example, if our array of strings contained the following:
|string list|
|-|
| '13k' |
| '546' |
| '9M' |
| '12M' |
| '900k' |

Our desired output would be: ['k', 'M']

import numpy as np

def getcharacters(string_list):
    characters = []

    for value in string_list:
        for character in value:
            if character.isalpha():
                characters.append(character)
    return np.unique(characters)

Let's apply the function we created to our dataset to see what string values are present:

unique_characters = getcharacters(pop_2017["2017"].values)
unique_characters

Great! we now know that we need to handle thousands ('k'), millions ('M'), and billions ('B'). 

Now we know we need to adjust our `string2num()` function that we created earlier to accommodate replacing the 'k', 'M' and 'B' characters.

def string2num(string):
    if "k" in string:
        number = float(string[:-1]) * 1000
    elif "M" in string:
        number = float(string[:-1]) * 1e6
    elif "B" in string:
        number = float(string[:-1]) * 1e9
    else:
        number = float(string)
    return number

# Test our output
print(string2num("2"))
print(string2num("2k"))
print(string2num("2M"))
print(string2num("2B"))

Just like we did for the case of income, let's apply our updated `string2num()` function to our population dataset. Since we're modifying our data, let's be sure to make a copy of it first.

pop = pop_2017.copy(deep=True)

pop["2017"] = pop_2017["2017"].apply(string2num)
pop

This looks good, let's double check our data types again to make sure everything was converted over to floats and make sure there were no `NaN` values introduced during processing:

pop.dtypes

pop.isna().values.any()

Lastly, let's rename the column header from '2017' to 'population' so that its column label is clear and descriptive for when we merge these data frames later.

pop = pop.rename(columns={"2017": "population"})
pop

Excellent! This is ready to save to file.

pop.to_csv('data/intermediate/pop.csv', index=False)

Great - let's collect these steps into a single function that we can use to reproduce this process later, as we've done ith emissions and income:

import pandas as pd

def process_population(infile,outfile):
    # Helper function to convert string data within the file into numerical data
    def string2num(string):
        if "k" in string:
            number = float(string[:-1]) * 1000
        elif "M" in string:
            number = float(string[:-1]) * 1e6
        elif "B" in string:
            number = float(string[:-1]) * 1e9
        else:
            number = float(string)
        return number
    
    # Read the file, select the columns of interest
    pop_all = pd.read_csv(infile)
    pop_2017 = pop_all[["country", "2017"]]
    
    # Convert the textual data in the dataset into numerical data
    pop = pop_2017.copy(deep=True)
    pop["2017"] = pop_2017["2017"].apply(string2num)
    
    # Rename the columns to make them more readable and save the output to file
    pop = pop.rename(columns={"2017": "population"})
    pop.to_csv(outfile, index=False)

# Test and make sure the file produces identical output as the original process:
raw_population_file = "data/raw/population_total.csv"
population_file = "data/intermediate/pop.csv"
process_population(raw_population_file,population_file)

With that done, we now have prepared our data for `co2`, `income`, and `pop`. All we need now is a way to identify the corresponding continent / region of the world for each country, and for that, we can look at our UN data.
