==> docs/_sources/notebooks/class_5/week_4/21_part5_merging.ipynb <==
# Merging the data

The goal for this part of the project is simple: merge the data into a single DataFrame.

We'll begin by loading and viewing each of our datasets to remind ourselves of their contents and to be able to compare them:

import pandas as pd

# Load the data
co2 = pd.read_csv('data/intermediate/co2.csv')
income = pd.read_csv('data/intermediate/income.csv')
pop = pd.read_csv('data/intermediate/pop.csv')
continent = pd.read_csv('data/intermediate/continent.csv')

co2

income

pop

continent

The first observation when looking across these four datasets is that they each have a different number of rows! 194, 195, 197, and 249, respectively. This means they don't all contain the same countries. But we need to combine them to plot them.

Think about what type of merge would be appropriate here? Left, right, inner, or outer? We only want to save the rows that are matching, so an inner merge seems like the right approach. Let's start by merging co2 and income based on the 'country' column:

data = pd.merge(co2, income, how="inner", on="country")
data

Excellent - we have 193 rows, which is as expected given the smallest size among the two dataframes being merged. Next lets add in population:

data = pd.merge(data, pop, how="inner", on="country")
data

And again we have 193 merged columns - looks like we have everything except for the continent labels added in. Adding in the continent labels is a little bit trickier, however, because if we look at the lists of the country/territory names in our `data` and `continent` frames, we can see that some of the names that are the same are simply written a bit differently. For example the following pairs represent the same territory:

- 'Brunei', 'Brunei Darussalam'
- 'Bolivia', 'Bolivia (Plurinational State of)'
- 'Venezuela','Venezuela (Bolivarian Republic of)'
- 'St. Vincent and the Grenadines', 'Saint Vincent and the Grenadines'
- 'Cote d'Ivoire', 'Côte d’Ivoire'

This means that these will not be merged together using the `merge()` method because their values are not identical.

We will create a method for identifying matching pairs of all names that we can confidently match, and will manually label the continent for the remaining territories. Looking at the data, it looks like we have 4 general categories:

1. **Exact matches**. Character-for-character equality. E.g., 'Burundi' <-> 'Burundi'
2. **Subset matches**. The "shorthand" name is a subset of the full name. E.g., 'Bolivia' <-> 'Bolivia (Plurinational State of)'
3. **Near matches**. The names are VERY similar, but a few characters may be different. E.g., 'Cape Verde' <-> 'Cabo Verde'
4. **Unmatched examples**. Cases that do not have corresponding pairs. E.g. 'Holy See'

Since the lists to be merged are of length 193 and 249, respectively, then a simple approach to merging them is to manually correct the names of countries/territories in the `continents` frame to match the rest. This approach is fine, however, this is not practical for larger lists which will commonly be encountered. Instead, we will find a way to minimize the need for manual intervention and identify the nearest match from one list to the other by considering the four possible cases above. As with many aspects of data science, we cannot always eliminate the need for some manual effort in preparing the data, but we will reduce it greatly.

We'll use the list of 4 categories above and test for each. What is important here is that order matters. If we perform a match, the matched strings will be removed from further consideration going forward so we don't create duplicate matches. Additionally, we'll match those quantities that most likely match first (e.g. exact matches before subset matches, before near matches). Since the lengths of the territory lists are 193 and 249, at most there will be 193 matches.

We'll be using the `difflib` package and the `SequenceMatcher` object and `get_close_matches` function. `SequenceMatcher` provides a similarity score for comparing two strings and takes on values between 0 and 1 where 1 is an exact match and 0 is no match at all. The `get_close_matches` function provides a list of the closest matches to the input string based on the `SequenceMatcher` similarity score.

We'll follow the pseudocode shared earlier for this:
1. Identify any exact matches and remove them from the list of strings to match
2. Identify subset matches and remove them from the list of strings to match
3. One at a time, create near matches based on the similarity score of each possible match
4. Repeat (3) until all 193 of the territories have been matched
5. Return the list of tuples with each match and its corresponding similarity score (set all exact matches and subset matches to a similarity score of 1)
6. Determine a threshold for the similarity score, below which assigning a match is not correct
7. Manually assign the content for the unmatched examples

from difflib import SequenceMatcher, get_close_matches

def similar(a, b):
    return SequenceMatcher(None, a, b).ratio()

def closest(a, B):
    return get_close_matches(a, list(B), n=1, cutoff=0)

long = list(continent["country"].values)
short = list(data["country"].values)
long_strings = long.copy()
short_strings = short.copy()

# First pass, match those with exact subsets
matches = []
for ss in short:
    candidates = []
    for ls in long:
        if ss in ls:
            candidates.append(ls)
    if len(candidates) > 0:
        selected = closest(ss, candidates)[0]
        sim = similar(ss, selected)
        long_strings.remove(selected)
        short_strings.remove(ss)
        matches.append([ss, selected, 1])

# Second pass, match what remains using proximity
slength = len(short_strings)
while slength > 0:
    matching_options = []
    similarity = []
    for ss in short_strings:
        selected = closest(ss, long_strings)[0]
        similarity.append(similar(ss, selected))
        matching_options.append([ss, selected])
    x = 1
    max_index = similarity.index(max(similarity))
    sim = max(similarity)
    long_strings.remove(matching_options[max_index][1])
    short_strings.remove(matching_options[max_index][0])
    matches.append([matching_options[max_index][0], matching_options[max_index][1], sim])
    slength = len(short_strings)

for e in matches:
    print(f"SS = {e[0]}, LS = {e[1]}, Sim = {e[2]}")

If we look at the full list, we can see that when the similarity score drops below 0.5, the match is typically incorrect, so we only replace those cases where the similarity score is greater than 0.5.

for match in matches:
    if match[2] > 0.5:
        continent = continent.replace(match[1], match[0])

Now we're ready to merge our data!

data = pd.merge(data, continent, how="left", on="country")
data

Since we had some unmatched cases, we have some values for `continent` that are `NaN`, let's identify those so we can manually assign a continent to each.

data[data.isnull().any(axis=1)]

Since we matched so many, we only have 6 remaining that are unmatched! That's a lot less work to manually assign. And let's remind ourselves of the list of continents to consider assigning:

continent["continent"].unique()

We can use the index above to manually assign each of those and complete our dataset! Let's see how it looks:

data.at[35, "continent"] = "Africa"
data.at[72, "continent"] = "Asia"
data.at[90, "continent"] = "Asia"
data.at[94, "continent"] = "Asia"
data.at[140, "continent"] = "Asia"
data.at[177, "continent"] = "Asia"
data

Let's double check for any issues including any null or Nan values:

data.isnull().any()

data[data.isna().any(axis=1)]

data[data.isnull().any(axis=1)]

Wonderful! We have our dataset successfully merged. 

## Changing the data from being divided into continents to 4 global regions

We just have one final data processing step before we create our plot. If you look at the chart on GapMinder's website, instead of using 7 continent categories, they use four categories for the world grouping Asia and Oceania into Asia. Let's do the same here by replacing any instances of "Oceania" with "Asia". Here we can use the `replace()` method.

data["continent"].unique()

data["continent"] = data["continent"].replace("Oceania", "Asia")
data["continent"].unique()

Phew! We have done a LOT of data processing to get ready to make our plot, which we'll do in the next lesson. Our data are the fuel for all the analyses we do in data science and often its takes some effort to prepare them for use. Having a clean dataset like we do now is entirely worth it as we can now explore these data by plotting them!

data.to_csv('data/intermediate/data.csv', index=False)

As a final step, let's collect all of our code for merging our data into a function we can rerun as needed to repeat this process:

import pandas as pd
from difflib import SequenceMatcher, get_close_matches

def match(continent, data):

    def similar(a, b):
        return SequenceMatcher(None, a, b).ratio()

    def closest(a, B):
        return get_close_matches(a, list(B), n=1, cutoff=0)

    long = list(continent["country"].values)
    short = list(data["country"].values)
    long_strings = long.copy()
    short_strings = short.copy()

    # First pass, match those with exact subsets
    matches = []
    for ss in short:
        candidates = []
        for ls in long:
            if ss in ls:
                candidates.append(ls)
        if len(candidates) > 0:
            selected = closest(ss, candidates)[0]
            sim = similar(ss, selected)
            long_strings.remove(selected)
            short_strings.remove(ss)
            matches.append([ss, selected, 1])

    # Second pass, match what remains using proximity
    slength = len(short_strings)
    while slength > 0:
        matching_options = []
        similarity = []
        for ss in short_strings:
            selected = closest(ss, long_strings)[0]
            similarity.append(similar(ss, selected))
            matching_options.append([ss, selected])
        x = 1
        max_index = similarity.index(max(similarity))
        sim = max(similarity)
        long_strings.remove(matching_options[max_index][1])
        short_strings.remove(matching_options[max_index][0])
        matches.append([matching_options[max_index][0], matching_options[max_index][1], sim])
        slength = len(short_strings)
    
    # Retain matches with a similarirty score greater than 0.5
    for match in matches:
        if match[2] > 0.5:
            continent = continent.replace(match[1], match[0])
    
    return continent


def merge_data(co2_file, income_file, pop_file, continent_file,outfile):
    # Load the data
    co2 = pd.read_csv(co2_file)
    income = pd.read_csv(income_file)
    pop = pd.read_csv(pop_file)
    continent = pd.read_csv(continent_file)
    
    # Merge CO2 and income:
    data = pd.merge(co2, income, how="inner", on="country")
    
    # Next merge in population:
    data = pd.merge(data, pop, how="inner", on="country")
    
    # Next align the continents data for merging, then merge it in:
    continent = match(continent, data)
    data = pd.merge(data, continent, how="left", on="country")
    
    # Manually adjust the unmatched continents:
    data.at[35, "continent"] = "Africa"
    data.at[72, "continent"] = "Asia"
    data.at[90, "continent"] = "Asia"
    data.at[94, "continent"] = "Asia"
    data.at[140, "continent"] = "Asia"
    data.at[177, "continent"] = "Asia"
    
    # Replace Oceania with Asia:
    data["continent"] = data["continent"].replace("Oceania", "Asia")
    
    # Save to file
    data.to_csv(outfile, index=False)
    
# Test and make sure the file produces identical output as the original process:
co2_file = 'data/intermediate/co2.csv'
income_file = 'data/intermediate/income.csv'
population_file = 'data/intermediate/pop.csv' 
continents_file = 'data/intermediate/continent.csv'
merged_data_file = 'data/intermediate/data.csv'
merge_data(co2_file, income_file, population_file, continents_file, merged_data_file)
