==> docs/_sources/notebooks/class_5/week_4/99_week4_wrap_up_quiz_w_solutions_code.ipynb <==
# Module 4 Wrap-up Quiz

One of the most important parts of this module was the data merging that you performed. This quiz will provide an opportunity to evaluate the merge that you just completed and to use the data to ask some additional questions about related to the data. Return to your merged data file and answer the following questions by querying the data.

## Q1. For countries with a population of 5 million or more, which country has the highest GDP per capita?

(Answer = Singapore)


import pandas as pd

df = pd.read_csv('data/intermediate/data.csv')
df_geq_5M = df.loc[df['population']>5e6]
df_sorted = df_geq_5M.sort_values('income', ascending=False)
df_sorted['country'].values[0]


## Q2. For the country you identified in Q1, what are it's carbon dioxide emissions per person in metric tonnes for the year provided in the data?

(Answer = 7.1)


df_sorted['co2'].values[0]


## Q3. For countries with population of 5 million or more, which country has the highest carbon dioxide emissions per person?

(Answer = United Arab Emirates)


df_sorted = df_geq_5M.sort_values('co2', ascending=False)
df_sorted['country'].values[0]

## Q4. For your answer in Q3, what is the average GDP per person in U.S. Dollars?

(Answer = 67000)

df_sorted['income'].values[0]


## Q5. For countries with a population of more than 10 million, which country has the highest CO2 emissions per person?

(Answer = Saudi Arabia)

df_geq_10M = df.loc[df['population']>10e6]
df_sorted_10M = df_geq_10M.sort_values('co2', ascending=False)
df_sorted_10M['country'].values[0]


## Q6. For your answer to Q5, what is the average GPD per person in U.S. Dollars?

(Answer = 47600)


df_sorted_10M['income'].values[0]

## Q7. Which of the four regions contains the highest number of unique countries in the dataset?

A. Africa
B. Americas
C. Asia
D. Europe

(Answer = C)


counted_df = df.groupby('continent').count()
print("Count by Group:")
print(counted_df['country'])



## Q8. How many countries are in the region in the answer to Q7?

(Answer = 64)

See above for how it was calculated.


## Q9. Which of the four regions has the smallest population?

A. Africa
B. Americas
C. Asia
D. Europe

(Answer = D)

summed_df = df.groupby('continent').sum()
print("Sum by Population:")
print(summed_df['population'])

## Q10. What is the total population of the region from Q9?

(Answer = 746185000)

d = summed_df['population']
d['Europe']
